+++ src/infrastructure/drizzle/schema/index.ts
@@ -1,4 +1,6 @@
 export * from './like.schema'
+export * from './quiz-result.schema'
+export * from './quiz-selection.schema'
 export * from './ranking.schema'
 export * from './text-to-speech.schema';
 export * from './user.schema'
